modified the file of config firebase

+++ Database/firebase-config.js
@@ -1,7 +1,7 @@
 
 import { initializeApp } from "https://www.gstatic.com/firebasejs/10.12.2/firebase-app.js";
 import { getFirestore, collection, addDoc, getDocs, getDoc, doc, setDoc, deleteDoc, updateDoc } from "https://www.gstatic.com/firebasejs/10.12.2/firebase-firestore.js";
-import { getAuth, createUserWithEmailAndPassword, signInWithEmailAndPassword, signOut } from "https://www.gstatic.com/firebasejs/10.12.2/firebase-auth.js";
+import { getAuth, createUserWithEmailAndPassword, signInWithEmailAndPassword, signOut, onAuthStateChanged} from "https://www.gstatic.com/firebasejs/10.12.2/firebase-auth.js";
 import { getStorage, ref, uploadBytes, getDownloadURL, deleteObject } from "https://www.gstatic.com/firebasejs/10.12.2/firebase-storage.js";
 import { getDatabase, set, push, onValue, update} from "https://www.gstatic.com/firebasejs/10.12.2/firebase-database.js";
 
@@ -21,4 +21,4 @@ const auth = getAuth(app);
 const storage = getStorage(app);
 const database = getDatabase(app);
 
-export { app, db, auth, storage, update,getDoc, deleteDoc, deleteObject, getDownloadURL, uploadBytes, signOut, database, set, push, onValue, createUserWithEmailAndPassword, signInWithEmailAndPassword, collection, addDoc, getDocs, setDoc, doc, ref, updateDoc };
+export { app, db, auth, storage, update,getDoc, deleteDoc, deleteObject, getDownloadURL, uploadBytes, signOut, database, set, push, onValue, createUserWithEmailAndPassword, signInWithEmailAndPassword, collection, addDoc, getDocs, setDoc, doc, ref, updateDoc, onAuthStateChanged, getAuth};
